+++ card/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CardConfig(AppConfig):
+    name = 'card'
